Fixes tests on Solaris
(backport <https://github.com/rust-lang/libc/pull/3864>)
(cherry picked from commit ec3c338269a499778796072088a1d5a0d9eb186b)
