+++ libc-test/semver/unix.txt
@@ -189,10 +189,6 @@ ISTRIP
 IXANY
 IXOFF
 IXON
-LOCK_EX
-LOCK_NB
-LOCK_SH
-LOCK_UN
 LOG_ALERT
 LOG_AUTH
 LOG_CONS
