+++ src/unix/solarish/x86_64.rs
@@ -1,3 +1,9 @@
+cfg_if! {
+    if #[cfg(target_os = "solaris")] {
+        use unix::solarish::solaris;
+    }
+}
+
 pub type greg_t = ::c_long;
 
 pub type Elf64_Addr = ::c_ulong;
@@ -46,6 +52,10 @@ s! {
         pub dlpi_phnum: ::Elf64_Half,
         pub dlpi_adds: ::c_ulonglong,
         pub dlpi_subs: ::c_ulonglong,
+        #[cfg(target_os = "solaris")]
+        pub dlpi_tls_modid: ::c_ulong,
+        #[cfg(target_os = "solaris")]
+        pub dlpi_tls_data: *mut ::c_void,
     }
 }
 
@@ -71,7 +81,18 @@ s_no_extra_traits! {
         pub uc_sigmask: ::sigset_t,
         pub uc_stack: ::stack_t,
         pub uc_mcontext: mcontext_t,
-        pub uc_filler: [::c_long; 5],
+        #[cfg(target_os = "illumos")]
+        pub uc_brand_data: [*mut ::c_void; 3],
+        #[cfg(target_os = "illumos")]
+        pub uc_xsave: ::c_long,
+        #[cfg(target_os = "illumos")]
+        pub uc_filler: ::c_long,
+        #[cfg(target_os = "solaris")]
+        pub uc_xrs: solaris::xrs_t,
+        #[cfg(target_os = "solaris")]
+        pub uc_lwpid: ::c_uint,
+        #[cfg(target_os = "solaris")]
+        pub uc_filler: [::c_long; 2],
     }
 }
